% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTiles.R
\name{print.sTiles}
\alias{print.sTiles}
\title{Print a sTiles factorization object}
\usage{
\method{print}{sTiles}(x, ...)
}
\arguments{
\item{x}{An "sTiles" object.}

\item{...}{Ignored, present for consistency with [print()].}
}
\value{
`x`, invisibly.
}
\description{
Print a sTiles factorization object
}
\examples{
if (sTiles_available()) {
Q <- Matrix::bandSparse(50, k = c(0, 1),
                        diagonals = list(rep(4, 50), rep(-1, 49)),
                        symmetric = TRUE)
  print(sTiles(Q))
}
}
