% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTiles.R
\name{print.sTiles_summary}
\alias{print.sTiles_summary}
\title{Print a sTiles summary}
\usage{
\method{print}{sTiles_summary}(x, ...)
}
\arguments{
\item{x}{An object of class "sTiles_summary" from [sTiles_summary()].}

\item{...}{Ignored, present for consistency with [print()].}
}
\value{
`x`, invisibly.
}
\description{
Print a sTiles summary
}
\examples{
if (sTiles_available()) {
Q <- Matrix::bandSparse(50, k = c(0, 1),
                        diagonals = list(rep(4, 50), rep(-1, 49)),
                        symmetric = TRUE)
  print(sTiles_summary(sTiles(Q)))
}
}
