% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTiles.R
\name{sTiles_library_path}
\alias{sTiles_library_path}
\title{Path of the loaded sTiles solver library}
\usage{
sTiles_library_path()
}
\value{
The absolute path of the loaded shared library, as a string.
}
\description{
Loads the solver if this session has not loaded it yet, then reports which
file answered. Useful when several builds are installed and you need to know
which one is in use.
}
\examples{
if (sTiles_available()) {
  sTiles_library_path()
}
}
\seealso{
[sTiles_available()], [sTiles_install_library()]
}
