% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTiles.R
\name{sTiles_logdet}
\alias{sTiles_logdet}
\title{Log-determinant of the factorized matrix}
\usage{
sTiles_logdet(x)
}
\arguments{
\item{x}{A factorized "sTiles" object.}
}
\value{
The log-determinant, a single number.
}
\description{
Returns `log(det(Q))`, computed from the factor as `2 * sum(log(diag(L)))`.
}
\examples{
if (sTiles_available()) {
Q <- Matrix::bandSparse(50, k = c(0, 1),
                        diagonals = list(rep(4, 50), rep(-1, 49)),
                        symmetric = TRUE)
  s <- sTiles(Q)
  sTiles_logdet(s)
  sTiles_close(s)
}
}
\seealso{
[sTiles()], [sTiles_summary()]
}
