% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sTiles.R
\name{sTiles_version}
\alias{sTiles_version}
\title{Version of the sTiles solver library}
\usage{
sTiles_version()
}
\value{
The solver's version string.
}
\description{
Version of the sTiles solver library
}
\examples{
if (sTiles_available()) {
  sTiles_version()
}
}
\seealso{
[sTiles_summary()]
}
